ajustes na organização das pastas, navegação e outras melhorias no fluxo
