+++ src/components/MiniProductType.js
