+++ src/routes/routes.js
@@ -1,22 +1,22 @@
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import { useAuth } from "../context/AuthContext";
 import { Ionicons } from "@expo/vector-icons";
 import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
-import SignInScreen from "../screens/SignInScreen";
-import SignUpScreen from "../screens/SignUpScreen";
 import { NavigationContainer } from "@react-navigation/native";
-import HomeScreen from "../screens/HomeScreen";
-import ProductScreen from "../screens/ProductScreen";
-import ProductDetailsScreen from "../screens/ProductDetailsScreen";
-import ShoppingCartScreen from "../screens/ShoppingCartScreen";
-import CheckoutScreen from "../screens/CheckoutScreen";
-import AccountScreen from "../screens/AccountScreen";
-import AccountDataScreen from "../screens/AccountDataScreen";
-import OrderScreen from "../screens/OrderScreen";
-import CreateProduct from "../screens/CreateProduct";
-import CategoryCreateScreen from "../screens/CategoryCreateScreen";
-import SportCreateScreen from "../screens/SportCreateScreen";
-import TeamCreateScreen from "../screens/TeamCreateScreen";
+import { useAuth } from "../contexts/AuthContext";
+import SignInScreen from "../screens/auth/SignInScreen";
+import SignUpScreen from "../screens/auth/SignUpScreen";
+import HomeScreen from "../screens/app/HomeScreen";
+import ProductScreen from "../screens/app/ProductScreen";
+import ProductDetailsScreen from "../screens/app/ProductDetailsScreen";
+import ShoppingCartScreen from "../screens/app/ShoppingCartScreen";
+import CheckoutScreen from "../screens/app/CheckoutScreen";
+import AccountScreen from "../screens/app/AccountScreen";
+import AccountDataScreen from "../screens/app/AccountDataScreen";
+import OrderScreen from "../screens/app/OrderScreen";
+import CreateProduct from "../screens/app/CreateProduct";
+import CategoryCreateScreen from "../screens/app/CategoryCreateScreen";
+import SportCreateScreen from "../screens/app/SportCreateScreen";
+import TeamCreateScreen from "../screens/app/TeamCreateScreen";
 
 const Stack = createNativeStackNavigator();
 const Tab = createBottomTabNavigator();
