+++ src/services/TeamService.js
@@ -1,8 +1,8 @@
-import { teamApi } from './api'
+import { gatewayApi } from './api'
 
 export async function getTeams() {
   try {
-    const response = await teamApi.get('/teams')
+    const response = await gatewayApi.get('/teams')
     return { teams: response.data, error: null }
   } catch (error) {
     return { error: error.message }
@@ -11,7 +11,7 @@ export async function getTeams() {
 
 export async function createTeam(team) {
   try {
-    const response = await teamApi.post('/teams/create', { name: team.nome })
+    const response = await gatewayApi.post('/teams/create', { name: team.nome })
     return { team: response.data, error: null }
   } catch (error) {
     return { error: error.message }
@@ -20,7 +20,7 @@ export async function createTeam(team) {
 
 export async function updateTeam(id, team) {
   try {
-    const response = await teamApi.put(`/teams/update/${id}`, { name: team.nome })
+    const response = await gatewayApi.put(`/teams/update/${id}`, { name: team.nome })
     return { team: response.data, error: null }
   } catch (error) {
     return { error: error.message }
@@ -29,7 +29,7 @@ export async function updateTeam(id, team) {
 
 export async function deleteTeam(id) {
   try {
-    await teamApi.delete(`/teams/delete/${id}`)
+    await gatewayApi.delete(`/teams/delete/${id}`)
     return { error: null }
   } catch (error) {
     return { error: error.message }
